requirements.txt
python-telegram-bot==20.4
torch==2.0.1
transformers==4.32.1
pandas==2.0.0
typer==0.9.0
tqdm==4.64.1 
scikit-learn==1.2.0
accelerate==0.22.0 